+++ src/libraries/Validations.sol
@@ -17,9 +17,9 @@ library Validations {
         address sender,
         address recipient,
         uint256 depositAmount,
-        uint64 startTime,
-        uint64 cliffTime,
-        uint64 stopTime
+        uint40 startTime,
+        uint40 cliffTime,
+        uint40 stopTime
     ) internal pure {
         // Checks: the common requirements for the `create` function arguments.
         _checkCreateArguments(sender, recipient, depositAmount);
@@ -40,10 +40,10 @@ library Validations {
         address sender,
         address recipient,
         uint256 depositAmount,
-        uint64 startTime,
+        uint40 startTime,
         uint256[] memory segmentAmounts,
         SD59x18[] memory segmentExponents,
-        uint64[] memory segmentMilestones,
+        uint40[] memory segmentMilestones,
         SD59x18 maxExponent,
         uint256 maxSegmentCount
     ) internal pure {
@@ -59,7 +59,7 @@ library Validations {
         uint256 segmentCount = segmentAmounts.length;
 
         // Imply the stop time from the last segment milestone.
-        uint64 stopTime;
+        uint40 stopTime;
         unchecked {
             stopTime = segmentMilestones[segmentCount - 1];
         }
@@ -112,10 +112,10 @@ library Validations {
     /// 4. The deposit amount is equal to the segment amounts summed up.
     function _checkSegments(
         uint256 depositAmount,
-        uint64 startTime,
+        uint40 startTime,
         uint256[] memory segmentAmounts,
         SD59x18[] memory segmentExponents,
-        uint64[] memory segmentMilestones,
+        uint40[] memory segmentMilestones,
         SD59x18 maxExponent,
         uint256 segmentCount
     ) private pure {
@@ -125,9 +125,9 @@ library Validations {
         }
 
         // Define the variables needed in the for loop below.
-        uint64 currentMilestone;
+        uint40 currentMilestone;
         SD59x18 exponent;
-        uint64 previousMilestone;
+        uint40 previousMilestone;
         uint256 segmentAmountsSum;
 
         // Iterate over the amounts, the exponents and the milestones.
