+++ test/integration/abstract-sablier-v2/AbstractSablierV2.t.sol
@@ -42,12 +42,12 @@ contract AbstractSablierV2 is SablierV2 {
         return address(0);
     }
 
-    function getStartTime(uint256 streamId) external pure override returns (uint64) {
+    function getStartTime(uint256 streamId) external pure override returns (uint40) {
         streamId;
         return 0;
     }
 
-    function getStopTime(uint256 streamId) external pure override returns (uint64) {
+    function getStopTime(uint256 streamId) external pure override returns (uint40) {
         streamId;
         return 0;
     }
